+++ main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS "functions.c" "app_main.c"
+idf_component_register(SRCS "ADC.c" "GPIO.c" "mcpwm.c" "functions.c" "app_main.c"
                        INCLUDE_DIRS "include""${CMAKE_SOURCE_DIR}""."
                        REQUIRES ssd1306 driver esp_adc hal esp_timer)
