+++ main/include/mcpwm.h
@@ -1,0 +1,7 @@
+#ifndef MCPWM_H
+#define MCPWM_H
+
+void mcpwm_init();
+void mcpwm_set_duty(float duty);
+
+#endif
