+++ sdkconfig.old
@@ -404,9 +404,9 @@ CONFIG_LIN_U_V_W_GPIO="25, 27, 12"
 #
 CONFIG_ENABLE_PWM=y
 CONFIG_TIMER_BASE_FREQ=40000000
-CONFIG_FREQ_PWM=40000
+CONFIG_FREQ_PWM=20000
 CONFIG_DUTY_PWM=50
-CONFIG_DEAD_TIME_PWM=850
+CONFIG_DEAD_TIME_PWM=500
 # end of PWM configuration
 
 #
@@ -1037,7 +1037,7 @@ CONFIG_ESP_WIFI_SOFTAP_SAE_SUPPORT=y
 CONFIG_ESP_WIFI_ENABLE_WPA3_OWE_STA=y
 # CONFIG_ESP_WIFI_SLP_IRAM_OPT is not set
 CONFIG_ESP_WIFI_STA_DISCONNECTED_PM_ENABLE=y
-CONFIG_ESP_WIFI_GMAC_SUPPORT=y
+# CONFIG_ESP_WIFI_GMAC_SUPPORT is not set
 CONFIG_ESP_WIFI_SOFTAP_SUPPORT=y
 # CONFIG_ESP_WIFI_SLP_BEACON_LOST_OPT is not set
 CONFIG_ESP_WIFI_ESPNOW_MAX_ENCRYPT_NUM=7
